Adding overlay to cuminc and refactor on scatter plot
