+++ client/termsetting/handlers/samplelst.js
@@ -181,6 +181,27 @@ export async function showTermsTree(div, callback, app, parentMenu, state = { tr
 		}
 	})
 }
+export function addPlotMenuItem(chartType, div, text, parentMenu, samplelstTW, id, parent) {
+	console.log(parent)
+	div
+		.append('div')
+		.attr('class', 'sja_menuoption sja_sharp_border')
+		//.html('Compare survival&nbsp;&nbsp;›')
+		.html(`${text}&nbsp;&nbsp;›`)
+		.on('click', e => {
+			const state = { tree: { usecase: { target: chartType, detail: 'term' } } }
+			if (chartType == 'survival') state.nav = { header_mode: 'hide_search' }
+			showTermsTree(
+				div,
+				term => {
+					openPlot(chartType, term, samplelstTW, parent.app, id, () => parent.newId)
+				},
+				parent.app,
+				parentMenu,
+				state
+			)
+		})
+}
 
 export function addMatrixMenuItems(menu, menuDiv, tw, app, id, state, newId) {
 	if (state.matrixplots) {
